+++ main.cpp
@@ -1,7 +1,7 @@
 
 #include "SimpleTCP.h"
 
-constexpr unsigned DEFAULT_PORT = 8080;
+constexpr uint16_t DEFAULT_PORT = 8080;
 
 int main() {
     SimpleTCP simpleTcp;
